models/resnet.py
import torch.nn as nn
from torchvision import models


def create_resnet18_model(num_of_classes):
    model = models.resnet18(pretrained=True)
    num_ftrs = model.fc.in_features
    model.fc = nn.Linear(num_ftrs, num_of_classes)

    return model


def create_resnet50_model(num_of_classes):
    model = models.resnet50(pretrained=True)
    num_ftrs = model.fc.in_features
    model.fc = nn.Linear(num_ftrs, num_of_classes)

    return model
